Remodel parser and tempload
