+++ restructuredpython/parser.py
@@ -1,5 +1,3 @@
-from .check_syntax import check_syntax
-import re
 # Copyright 2025 Rihaan Meher
 
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -17,8 +15,53 @@ import re
 from .check_syntax import check_syntax
 import re
 
+def wrap_loops_for_optimization(code):
+    """
+    Rewrites for/while loops with <OPTIMIZE ...> annotations into runtime functions
+    with decorators.
+    """
+    lines = code.splitlines()
+    modified_lines = []
+    i = 0
+    loop_counter = 0
+
+    while i < len(lines):
+        line = lines[i].strip()
+        if line.startswith("@optimize_loop("):
+            decorator_line = lines[i]
+            loop_line = lines[i + 1].strip()
+            loop_indent = len(lines[i + 1]) - len(loop_line)
+            func_name = f"_repy_optimized_loop_{loop_counter}"
+            loop_counter += 1
+
+            modified_lines.append(decorator_line)
+            modified_lines.append(" " * loop_indent + f"def {func_name}():")
+            modified_lines.append(" " * (loop_indent + 4) + loop_line)
+            i += 2
+
+            # Copy indented body lines until dedent or EOF
+            while i < len(lines):
+                body_line = lines[i]
+                if body_line.strip() == "":
+                    modified_lines.append(body_line)
+                    i += 1
+                    continue
+                body_indent = len(body_line) - len(body_line.lstrip())
+                if body_indent <= loop_indent:
+                    break
+                modified_lines.append(" " * (loop_indent + 8) + body_line.strip())
+                i += 1
+
+            modified_lines.append(" " * loop_indent + f"{func_name}()")
+        else:
+            modified_lines.append(lines[i])
+            i += 1
+
+    return "\n".join(modified_lines)
+
 def parse_repython(code, mode="classic"):
     """Parses reStructuredPython and converts it to Python."""
+    required_imports = set()
     def chain_pipeline(code):
         parts = [part.strip() for part in code.split('|>')]
         if len(parts) > 1:
@@ -82,9 +125,11 @@ def parse_repython(code, mode="classic"):
         if pending_optimize:
             if re.match(r'^\s*(for|while)\s+.*\{', processed_line):
                 modified_code.append(f"@optimize_loop({pending_optimize})")
+                required_imports.add("optimize_loop")
                 pending_optimize = None
             elif re.match(r'^\s*def\s+.*\{', processed_line):
                 modified_code.append(f"@optimize_function({pending_optimize})")
+                required_imports.add("optimize_function")
                 pending_optimize = None
 
         # Block conversions
@@ -106,4 +151,17 @@ def parse_repython(code, mode="classic"):
         else:
             modified_code.append(processed_line)
 
-    return '\n'.join(modified_code)
+    # Insert necessary imports at the top
+    import_lines = []
+    if "optimize_loop" in required_imports or "optimize_function" in required_imports:
+        imports = []
+        if "optimize_function" in required_imports:
+            imports.append("optimize_function")
+        if "optimize_loop" in required_imports:
+            imports.append("optimize_loop")
+        import_lines.append(f"from restructuredpython.predefined.subinterpreter import {', '.join(imports)}")
+
+    raw_code = '\n'.join(import_lines + modified_code)
+    final_code = wrap_loops_for_optimization(raw_code)
+    return final_code
+
