+++ restructuredpython/predefined/subinterpreter/__init__.py
@@ -1,0 +1,9 @@
+from .optimize import (
+    optimize_function,
+    optimize_loop
+)
+
+__all__ = [
+    "optimize_function",
+    "optimize_loop"
+]
